Removed image link from hmMacandCheese.html
The wrong file type was used to try and create this image. It used a
webp file type, which isn't recognized on the internet, at least not
by default. The four image types we should use are JPG, GIF, PNG, SVG.

If it isn't one of these four, dodge it for now. We don't know whats
up with that filetype.