New file jigld.html added, description added
We have added jigld.html file, which stands for Jigidy's Lazy Day. I
added a description for the recipe. Basically, it is a recipe that
is completely meant for the giggles so don't take it too seriously,
but I was reading another guys code for the recipe and just concluded
that I actually could make this however I wanted, so humour it is.