Added jigld file and linked it, final goals in README
I have added and linked the jigld file in order to provide a picture.

I have also added several final details to the README.md file that
serve as my final goals in this project. We are essentially going to
work to complete these objectives, and after that we'll check to make
sure the project is done to requirements and then publish it.