Images work, but need more. Also, background color
The images work. They're positioned correctly and are now responded to
CSS code, but there are some kinks to sort out with their dimensions
which may need more CSS selectors to sort out.

As a final spin to this project, I want to see if I can get some
background colors working for the page, on a gradient as well. Seems
like a way to push a little outside the bounds of what I've been
asked to do, so I can make further progress.